Task_3_FinGPT_Powered_Agents_MultiModal_Financial_Data/pyproject.toml
[project]
name = "task-3-fingpt-powered-agents-multimodal-financial-data"
version = "0.1.0"
description = "FinGPT-Powered Agents for MultiModal Financial Data - SecureFinAI Contest 2025 Task 3"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.4.0",
    "numpy>=1.21.0",
    "Pillow>=8.0.0",
    "tqdm>=4.60.0",
    "datasets>=2.0.0",
    "evaluate>=0.4.0",
    "pytesseract>=0.3.10",
    "easyocr>=1.6.0",
    "paddlepaddle>=2.4.0",
    "paddleocr>=2.6.0",
    "opencv-python>=4.5.0",
    "scikit-image>=0.19.0",
    "torch>=1.12.0",
    "transformers>=4.20.0",
    "beautifulsoup4>=4.10.0",
    "lxml>=4.6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
